Renamed the dialog service and component to .ng1
