+++ source/components/messageLog/messageLog.directive.tests.ts
@@ -15,7 +15,7 @@ import { defaultThemeValueName } from '../componentsDefaultTheme';
 import * as angular from 'angular';
 import 'angular-mocks';
 
-import { serviceName as dialogServiceName, IAutosaveDialogSettings } from '../../services/dialog/dialog.service';
+import { serviceName as dialogServiceName, IAutosaveDialogSettings } from '../../services/dialog/dialog.service.ng1';
 
 interface IMockMessageLogService {
 	visibleMessages: number[];
